annalib::threads

Type Alias UserThread

pub type UserThread = Thread;

UserThread extends Thread with response address methods

Aliased Type

pub struct UserThread { /* private fields */ }

Implementations

impl UserThread

pub fn new(ip: &Address, tid: usize) -> Self

create a new UserThread from the Address and tidthread id

pub fn with_offset(ip: &Address, tid: usize, base_offset: usize) -> Self

create a new UserThread with a port base offset

pub fn response_bind_address(&self) -> Address

return the ZMQ Address to use to bind to a response

pub fn response_connect_address(&self) -> Address

return the ZMQ Address to use to connect to a response